Add select field for label attribute in TextEditor
This adds an AttributeCombo component to the TextEditor in order to let
the user select the field / attribute which is used as label in the map.
Herewith the SymbolizerPreview is changed that it accepts the full
internal data object (instead of just the features), so the schema can
be forwarded to the new AttributeCombo.