Replace LocaleReceiver with HOC localize
Instead of wrapping every (internationalisation dependent) component with antd's LocaleReceiver and shifting the original render-function into a helper function, now, a HOC localize can be used. The render-function can be used as before (and as it should be used). To use internationalisation in a component, function localize from module LocaleWrapper must be called before exporting respective component.