Handle TextSymbolizers
Now, TextSymbolizerse will be handled properly. Openlayers only accepts an array of styles or a single style-function. But not a mixed array with styles and functions. Since TextSymbolizer returns a function, as soon as a TextSymbolizer is being used, all styles need to be wrapped into a function that returns an array of styles. Thus, we give openlayers only a single function.