Fix internationalisation minify bug
localize from LocaleWrapper.ts did not work properly when minified. Due to renaming of class names, localize could not find the matching properties in the language packs. Bug was fixed through adding the original class name of a component as additional parameter in localize.

Every class that is wrapped in localize now has a static property componentName. This property will then be added as parameter to localize.