breaking: force major release
BREAKING CHANGE: This forces a major release for a previously wrongly
formatted commit message (commit sha 30e19de16a6310dcd79bf4e4292b33bf80004c0d)
that lead to the unintended release of version 10.3.1, which includes a
breaking change. The code of this release will be identical with the code of
version 10.3.1.