feat: use GeoStylerContext composition on LineEditor (#2123)
BREAKING CHANGE: This removes the support of the deprecated
CompositionContext in favor of the new GeoStylerContext for the
LineEditor. This also removes the `defaultValue` property from
LineEditor. Please use GeoStylerContext.composition instead. For now,
composition of graphicFill and graphicStroke was changed without
replacement. Support will be re-added again in the future.